+++ R/get_schema.R
@@ -26,7 +26,7 @@
 #' @examples
 #' conn <- get_connection(drv = RSQLite::SQLite())
 #'
-#' dplyr::copy_to(conn, mtcars, name = "mtcars")
+#' dplyr::copy_to(conn, mtcars, name = "mtcars", temporary = FALSE)
 #'
 #' get_schema(conn)
 #' get_schema(get_table(conn, id("mtcars", conn = conn)))